+++ muon.h
@@ -6,14 +6,14 @@
 class Muon : public Lepton {
 	public :
 		int idbits;
-		int nMatches;
-		int validMuonHits;
+		float nMatches;
+		float validMuonHits;
 
 		Muon( float px_, float py_, float pz_, float en_, float ptErr_, float ecalIso_, float hcalIso_,
 				float trkIso_, float gIso_, float chIso_, float puchIso_, float nhIso_, int id_, int genid_,
 				float ensf_, float ensferr_, float d0_, float dZ_, float trkpt_, float trketa_, float trkphi_,
 				float trkchi2_, float trkValidPixelHits_, float trkValidTrackerHits_, float trkLostInnerHits_,
-				int idbits_, int nMatches_, int validMuonHits_ );
+				int idbits_, float nMatches_, float validMuonHits_ );
 
 		bool isTrackerMuon() const;
 		bool isGlobalMuon() const;
